Updated openstack/openstack Project: openstack-infra/project-config  d632bc4eec1ed156d1dedece31626729974dbab0
Import ansible-role-logrotate

We'll be using the role as part of windmill, currently we are not big
tent but we do want IRC notification for #openstack-windmill.  We also
add some -nv jobs as the first task after importing is to ensure jobs
are passing correctly. Then we'll make them vote properly.

Change-Id: I50f51f1a34d77e251053330f84c6f9f23bd38ed0
Signed-off-by: Paul Belanger <pabelanger@redhat.com>