Updated openstack/openstack Project: openstack/cinder  07a90676f79935507487da247cbc77f6542cbd99
Update OVO devref

On commit f0d34b7d9ba79af1b50c1e89e2cd3493075c2754 we changed the way we
did backports for OVOs during upgrades removing the need to bump objects
when related objects are modified.

This patch updates the documentation to reflect this as well as adds a
new example of backporting code for the case where we have added new
fields in the latest version of an object.

TrivialFix

Change-Id: I327235fdafdcee2d36395c4c64372caf6b79d1e2