Updated openstack/openstack Project: openstack/designate  5ccf2c38f12e3b57028906f88abe1091fcaa0db3
Fix test notification driver configuration

The tests try to configure oslo messaging notifications to use the
"test" driver, but it needs a list, not a single value. You can see the
warning in the logs:

WARNING [stevedore.named] Could not load s, e, t

This properly declare the driver as a single item list.

Change-Id: Ia0b57580c00fe0f7b9c0a19ac5190282972f624f