Updated openstack/openstack Project: openstack/horizon  7a5b4a71a31c132d515444aa699cc19ce2df9858
Make the "Select Boot Source" order deterministic

This patch makes the "Select Boot Source" dropdown in the Source step of
the new Launch Instance have a deterministic order.

The slight downside to this is that the order needs to be recalculated
as new possible boot sources are loaded. However, given that this is the
second step of the workflow, this is unlikely to ever be noticed by a
user.

Change-Id: Iaa2e37bae02c9ce5ab654ac8a3327c0834e7b49e
Closes-Bug: 1640493