Updated openstack/openstack Project: openstack/nova  387a66d9b9e2ba54810912ff0db0b340031e4fab
Add recreate functional test for regression bug 1689692

When paging through instances, if the marker is found in cell0
and there are more instances under the limit, we continue paging
through the cell(s) to fill the limit. However, since the marker
was found in cell0 it's not going to be in any other cell database
so we'll end up failing with a marker not found error.

This change adds a functional recreate test for the bug.

The fix will build on this to show when the bug is fixed and the
test will be changed to assert expected normal behavior.

Change-Id: I234e0425e7e800b32cea78f5c1d99997bc03343f
Partial-Bug: #1689692