Updated openstack/openstack Project: openstack/oslo.messaging  40532a2a457eba1179183654079e0ed81ddbcc48
Mark the Pika driver as deprecated

The Pika driver was intended to be a more stable and performant
replacement for the default rabbit driver.  However due to lack of
both maintainers and compelling evidence that pika is superior to the
existing rabbit driver in either performance or stability it has been
deprecated for removal.

See:
http://lists.openstack.org/pipermail/openstack-dev/2017-May/116679.html

Change-Id: I98e0123edd3248be665325833283689fc3a897f7