Updated openstack/openstack Project: openstack-infra/project-config  ea4ad175cfd09abb9c5b077934cb49a1f11f8d12
Fix puppet install and execution issues

When adding a repository in Ubuntu, one must run apt-get update prior
to trying to install a package from the new repository. Additionally
when this new package installs binaries in an alternative path, you
need to ensure the PATH is properly updated for it to work.

Change-Id: I285baa486221e496baa8e205bd7a92c33d07dd2f