Updated openstack/openstack Project: openstack/nova  bb4205e86341594f2fb5c07407ff74f649cea810
XenAPI: nova-compute cannot restart after manually delete VM

When a VM is deleted accidentally (e.g. hardware problem or manually
through the hypervisor rather than Nova), there is a mis-match of
information where the VM is still in nova DB but not the hypervisor.
If we start nova-compute service in this setup, it will fail due to an
untrapped exception when plugging VIFs.

Return an expected exception when Nova cannot find VM via xapi.

Closes-bug: #1693147

Change-Id: I937f5c202c9a4892e8aa56f74fad125791809f8c