Updated openstack/openstack Project: openstack/cinder  ae3ce62ae4be2ce414d188146ea3bc418aa2e548
Do proper cleanup if connect volume fails

If during the create image from volume the connect method fails with an
unexpected exception we will not be terminating the connection or
removing the export.  We will also not remove the export if the DB fails
when initializing the connection.

This patch fixes this making sure we are properly cleaning things up.

Change-Id: Ifa20366e6197bdce014d7f922d4700b684be1c23
Closes-Bug: #1685843