Updated openstack/openstack Project: openstack/trove  adff3563155231af72ae0166e6a00fa5980e7b64
Add port 16379 to conf.sample

This is a point that is easily hidden. In the case of not open
port 16379, redis cluster can still be deployed successfully,
but this success is a false success, there is no data communication
between the various nodes. This is because 16379 is the port for
data port[1]. We set the default value in cfg, but in the
case of conf,it is easy to be covered. Adding 16379 to conf.sample
helps the user to reduce some problem caused by the redis
configuration.

[1]:https://redis.io/topics/cluster-tutorial

Change-Id: If517072c1c875df68106af14dac1802bb959d17e