Updated openstack/openstack Project: openstack-infra/storyboard  bf30de7a999f4a5b6ea849f7084ff078043b64c4
Allow stories and tasks to be filtered by boards/worklists containing them

Currently this is only supported for the internal database API surface,
with the intention of using it to reduce the amount of work done by the
boards and worklists APIs.

This patch also names some aliases we create to make debugging less
painful.

It also modifies the join to the StorySummary query such that the
column name is no longer hardcoded. This handles the case when we
have filtered out private worklists from the result (when joining
to worklists), which results in an anonymous column name we cannot
hardcode.

Change-Id: I3aaa874ae623a44edd668471794f8146c6dabda7