Updated openstack/openstack Project: openstack/glance-specs  77a7519479760c48081a3dc9da4184e2a9c5da4f
Introduce 'untargeted' specs

This commit adds an 'untargeted' directory to hold orphaned specs
that were approved but aren't being worked on.  A new index.rst is
added explaining the idea, and adjustments are made to other
appropriate index pages.  Additionally, one glanceclient spec from
liberty was stuck in a non-release-name directory; it was moved to
the liberty directory, which in turn was slightly reorganized to
accommodate both glance and glanceclient specs.

Change-Id: I074fc782eca1c3e3c9a1c504501973300741b110