Updated openstack/openstack Project: openstack/horizon  117cec9b544c49d429499dd99ab9824ade33bf4f
doc: Ignore warnings from dashboard build in the doc build

There is no need to catch warnings from openstack_dashboard.settings
in the doc build job. This might be a bug of a recent Sphinx 1.6.x,
but I think it is useful to have this setting to avoid unexpected failures.

Change-Id: I9894622ba0f5fbb74e9dac016ed76ac9e3c2f733