Updated openstack/openstack Project: openstack/diskimage-builder  b708918b85208ce7741209c1ff5e969dc49b0882
Remove 'state' argument from later cmd_* calls

With I468dbf5134947629f125504513703d6f2cdace59 each node has a
reference to the global state object.  This means it gets pickled into
the node-list, which is loaded for later calls.  There is no need to
reload the state.json it and pass it for later cmd_* calls, as the
nodes can see it via the unpickled self.state

Change-Id: I9e2f8910f17599d92ee33e7df8e36d8ed4d44575