Updated openstack/openstack Project: openstack-infra/system-config  b3424363cdf41c3c722811544af0ca1953f4d0fa
Bump virtualenv to 15.1.0

Due to a bad puppet-pip patch that I have merged, it was possible for
our virtualenv python installs to be switch to python3.  As a result,
now we have a mix of virtualenv versions under python2 and python3 for
our control plane.

As a result, bump virtualenv to the latest version so we can reset the
base versions across all our servers.

Change-Id: I9f2819b697dcffddc9ca7c06bfcf72766ec86d40
Signed-off-by: Paul Belanger <pabelanger@redhat.com>