Updated openstack/openstack Project: openstack/requirements  7210676bc1bf35dd6258b269c1d546ed7293416b
remove cap on pysnmp from global-requirements.txt

The cap was added in a time it was considered the norm four years ago.
The cap has never restricted anything (the current version in
upper-constraints.txt is the latest), so let's bring it in line with
the general policy of not having caps.

Change-Id: Ie804bc9d758a5f19d2bfc9e02cabe1bd0a2ac8a9