Updated openstack/openstack Project: openstack/requirements  a1bd9004d67cbb55070e2caf987744cb31c468b9
Blacklist oslo.messaging 5.25.0

The default executor of o.m is 'blocking', but after change [1],
o.m with blocking is broken. 'eventlet' and 'non-blocking' work
fine and most services use 'eventlet', so this is why it was
noticed by magnum which used the default one.

Blacklist 5.25.0 until the 'blocking' executors is fixed.

[1] https://review.openstack.org/#/c/463673/

Change-Id: I9943a6f69dc7438364004e5604bafc7d4646b1c1
Related-Bug: #1694728