Updated openstack/openstack Project: openstack/governance  22a13dc3d87f76b3f48946a6d299f5a18ce64b13
Add ansible-hardening to OpenStack-Ansible

After a discussion on the mailing list[0], the general consensus is
that the 'openstack-ansible-security' project should be renamed to
'ansible-hardening' to better reflect the goals of the project. It
can be used in conjunction with OpenStack-Ansible, and without.
Having 'openstack-ansible-' in the name led to confusion and many
users felt like it could only be used with an OpenStack-Ansible
deployment.

A simple repo rename isn't sufficient due to legacy codebases
referring to git.openstack.org. See the project-config[1] patch for
more details.

[0] http://lists.openstack.org/pipermail/openstack-dev/2017-May/116922.html
[1] https://review.openstack.org/467339

Depends-On: I208d2d7ed125264efa4266b8927810268345b8ce
Change-Id: Id533715a62c2fc5561cae0a8d7e7f5d3b0caa3cc