Updated openstack/openstack Project: openstack/governance  1ec76ea49433487edd625002507a6e00ff0eb355
Add ldappool to the keystone managed projects

The ldappool project was adopted by several memebers of the keystone
development team. This commit formally adds the project to the list
of keystone-managed openstack projects.

This will allow us to manage releases the same way we do with the
other libraries we manage.

Change-Id: Ic7b7b693dfb36ca29a8499e2bb3956145ec3c098