Updated openstack/openstack Project: openstack/governance  81ee96e95f67f71c5c57d330bb608e1206587d42
Document voting process for `formal-vote` patches

A while ago it was decided[0] that motions, before being
merged/approved, should be discussed and voted on during the Technical
Committee meetings. This worked well back in the days but it's not
necessary anymore. This was carried over even though we now have an
asynchronous voting system and it was never properly documented or
re-evaluated.

This patch documents how formal-vote patches are voted on and eventually
approved.

[0] http://lists.openstack.org/pipermail/openstack-dev/2013-August/013339.html

Change-Id: Ie84bacb9038d82bdfb518724a8e6fb726f5a0316