Updated openstack/openstack Project: openstack-infra/project-config  2f250eb52dee6f79816b2f7de7ab69c426f91e6a
Temporarily disable Beaker jobs for puppet-midonet

Some Puppet modules were not pinned in our Puppetfile and are now
breaking every patch that we submit due to the fact that some modules
are Puppet 4 only.

We need to disable beaker temporarily, then pin versions on the
Puppetfile and finally restore beaker.

Change-Id: I7734202254f4fe88af63f88be9db4c66ead26da6
Signed-off-by: Alejandro Andreu <alejandroandreu@openmailbox.org>