Updated openstack/openstack Project: openstack-infra/shade  977fb73722c7b4a41994c0a7b339906167b85c3e
Set some logger names explicitly

In the next patch we'll add documentation about what shade logs to what
logging facilities. To make sure that those named facilities are
maintained as part of the interface a consumer can count on, set
non-root logger names explicitly.

Also, this changes the logger name for the "waiting" messages from
shade._utils to shade.iterate_timeout. _utils is an implemenation detail
and such subject to change. The logger for it is not.

Change-Id: I2ff413062c4cc3b49872fa8fd24e5daade063187