Updated openstack/openstack Project: openstack-infra/nodepool  9a15144c9bce987009f4125f2f124f15de252bac
Consider hostname in determining local builds

If DIB fails to produce local files, or if we remove a failing diskimage
from the config before the ZooKeeper node has been removed, we never
remove the znode. Work around this by considering any recorded builds
that have a 'builder' that matches the host doing the cleanup as owned
by that host.

Change-Id: Id3b1e2ae0d0f71209cd66091a7d69143b0b4b20c