Updated openstack/openstack Project: openstack-infra/elastic-recheck  0e0ec10c7809ab919743dd9764d9cd903eec1ed9
Query for HOST_IP not found issue

This is a query for bug 1693689 that will track instances where devstack
jobs fail because the host ip couldn't be determined. Underlying issue
is jobs not using the safe ip address range defined in d-g which
occasionally leads to address range conflicts with the underlying cloud
instances.

Change-Id: I54d34db640d6ca6c7c39783bba8246fb5e951467
Related-bug: 1693689